equation_solver_service/requirements.txt
psycopg2==2.9.3
django==4.1.2
djangorestframework
requests
sympy
scipy
django-rest-swagger
drf_yasg
django-cors-headers
google-generativeai
matplotlib